+++ PoliticPolls.DataModel/Entities/PoliticianSets.cs
@@ -5,9 +5,9 @@ namespace PoliticPolls.DataModel
 {
     public partial class PoliticianSets
     {
-        public decimal IdPoll { get; set; }
-        public decimal IdPolitician { get; set; }
-        public decimal? Rating { get; set; }
+        public int IdPoll { get; set; }
+        public int IdPolitician { get; set; }
+        public int? Rating { get; set; }
 
         public Politicians Politician { get; set; }
         public Poll Poll { get; set; }
